+++ include/orderbook.h
@@ -15,7 +15,7 @@ using OrderPtr = std::shared_ptr<Order>;
 struct Trade {
     uint64_t buy_order_id;
     uint64_t sell_order_id;
-    double price;
+    int64_t price;
     uint64_t quantity;
     uint64_t timestamp;
 };
@@ -23,10 +23,10 @@ struct Trade {
 class OrderBook {
 private:
     // Bids: price -> queue of orders (descending by price)
-    std::map<double, std::queue<OrderPtr>, std::greater<double>> bids;
+    std::map<int64_t, std::queue<OrderPtr>, std::greater<double>> bids;
     
     // Asks: price -> queue of orders (ascending by price)
-    std::map<double, std::queue<OrderPtr>, std::less<double>> asks;
+    std::map<int64_t, std::queue<OrderPtr>, std::less<double>> asks;
     
     // Track all active orders by ID for cancellation/modification
     std::unordered_map<uint64_t, OrderPtr> active_orders;
@@ -44,7 +44,7 @@ public:
     
     // Order management
     bool cancelOrder(uint64_t order_id);
-    bool modifyOrder(uint64_t order_id, double new_price, uint64_t new_quantity);
+    bool modifyOrder(uint64_t order_id, int64_t new_price, uint64_t new_quantity);
     
     double getBestBid() const;
     double getBestAsk() const;
